requirements.txt
langchain
qdrant-client
sentence-transformers
streamlit
huggingface_hub
langchain-community
langchain-huggingface
libmagic
unstructured[md]
uv
python-dotenv
presidio-analyzer
presidio-anonymizer
llama-cpp-python
flask
faiss-cpu
tiktoken
